add getAssetIssueList
